# docs/swagger.yaml
basePath: /
definitions:
...
  models.Answer:
    properties:
      answer_text:
        type: string
      answer_value:
        type: integer
      question_id:
        type: integer
    type: object
...
  models.MatchedProfile:
    properties:
      alcohol:
        type: string
      bio:
        type: string
      birth_place:
        type: string
      birth_time:
        type: string
      caste:
        type: string
      city:
        type: string
      civil_status:
        type: string
      country_code:
        type: string
      created_at:
        type: string
      date_of_birth:
        type: string
      dietary_preference:
        type: string
      district:
        type: string
      employment_status:
        type: string
      father_occupation:
        type: string
      field_of_study:
        type: string
      gender:
        type: string
      height_cm:
        type: integer
      highest_education:
        type: string
      horoscope:
        type: string
      horoscope_available:
        type: boolean
      id:
        type: integer
      institution:
        type: string
      interests:
        description: Array of interests
        items:
          type: string
        type: array
      languages:
        items:
          type: string
        type: array
      last_active_at:
        type: string
      location:
        type: string
      metadata:
        type: string
      moderation_status:
        type: string
      mother_occupation:
        type: string
      nakshatra:
        type: string
      occupation:
        type: string
      postal_code:
        type: string
      profile_image_thumb_url:
        type: string
      profile_image_url:
        type: string
      province:
        type: string
      reasons:
        type: object
      religion:
        type: string
      religion_detail:
        type: string
      score:
        type: number
      siblings:
        type: string
      siblings_count:
        type: integer
      sinhala_raasi:
        type: string
      smoking:
        type: string
      updated_at:
        type: string
      user_id:
        description: Foreign key to users table
        type: integer
      username:
        type: string
      verified:
        type: boolean
      weight_kg:
        type: integer
    type: object
...
  models.Question:
    properties:
      id:
        type: integer
      options:
        items:
          type: string
        type: array
      questionText:
        type: string
      questionType:
        type: string
    type: object
...
  utils.AnswersResponse:
    properties:
      answers:
        items:
          $ref: '#/definitions/models.Answer'
        type: array
    type: object
  utils.ErrorResponse:
    properties:
      details:
        type: string
      error:
        type: string
    type: object
...
  utils.MessageResponse:
    properties:
      message:
        type: string
    type: object
  utils.QuestionnaireResponse:
    properties:
      questions:
        items:
          $ref: '#/definitions/models.Question'
        type: array
    type: object
...
info:
  contact: {}
  description: API documentation for the Dating App backend.
  title: Dating App API
  version: "1.0"
paths:
...
  /user/matches/{user_id}:
    get:
      description: Combines compatibility scores from the matching service with profile
        details.
      parameters:
      - description: User ID
        in: path
        name: user_id
        required: true
        type: integer
      - description: Optional limit for number of matches
        in: query
        name: limit
        type: integer
      - description: Optional offset for pagination
        in: query
        name: offset
        type: integer
      - description: Minimum score filter
        in: query
        name: minScore
        type: number
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.MatchedProfile'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Retrieve the best matches for a user
      tags:
      - Matches
...
  /user/questionnaire:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.QuestionnaireResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Retrieve matchmaking questionnaire
      tags:
      - Questionnaire
  /user/questionnaireAnswers:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.AnswersResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/utils.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Retrieve questionnaire answers for the authenticated user
      tags:
      - Questionnaire
...
  /user/submitQuestionnaire:
    post:
      consumes:
      - application/json
      parameters:
      - description: Questionnaire answer
        in: body
        name: answer
        required: true
        schema:
          $ref: '#/definitions/models.Answer'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.MessageResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/utils.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Submit questionnaire answers
      tags:
      - Questionnaire
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"